requirements.txt
Django==4.2.7
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.4
django-jazzmin==2.6.0
django-mptt==0.15.0
django-parler==2.3
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
Markdown==3.5.1
moviepy==1.0.3
phonenumbers==8.13.26
Pillow==9.5.0
psycopg2==2.9.9
requests==2.31.0
